Move cluster functions to helpers.py
Not proper utils, and not pandas-dependent.
